requirements.txt
google-cloud-tpu
rich
fire